+++ src/main/java/staysplit/hotel_reservation/hotel/dto/request/DeleteHotelRequest.java
@@ -1,5 +1,5 @@
 package staysplit.hotel_reservation.hotel.dto.request;
 
 public record DeleteHotelRequest(
-        Long hotelId
+        Integer hotelId
 ) {}
