+++ src/main/java/staysplit/hotel_reservation/hotel/dto/response/DeleteHotelResponse.java
@@ -2,5 +2,5 @@ package staysplit.hotel_reservation.hotel.dto.response;
 
 public record DeleteHotelResponse(
         String message,
-        Long hotelId) {
+        Integer hotelId) {
 }
