+++ src/main/java/staysplit/hotel_reservation/room/domain/dto/response/RoomDeleteResponse.java
@@ -1,7 +1,7 @@
 package staysplit.hotel_reservation.room.domain.dto.response;
 
 public record RoomDeleteResponse(
-        Long hotelId,
-        Long roomId
+        Integer hotelId,
+        Integer roomId
 ) {
 }
